fix. mv resHandle.js to service folder, rename post routes controller name, and add error handle for updatePost
